refactor: extract duplicate media helper methods to shared MediaHelper class
